refactoring code into seperate files
